+++ src/assets/products.json
@@ -125,7 +125,7 @@
       },
       "fields": {
         "title": "My One",
-        "price": 999999.99,
+        "price": 499.99,
         "description": "this one is raised by me on my apertment.One day I saw that she was starting to grow leaves So I put it in a water, this one is special to me, this is why it a bit expensive. ",
         "image": {
           "fields": {
