--- src/modules/Catalog.js
@@ -1,14 +1,0 @@
-// import Utilities from "../utils/Utilities";
-// import Product from "./Product";
-// class Catalog {
-
-//     constructor() {
-//         const productQuant = Utilities.randomInteger(6, 24)
-//         this.productsArray = new Array(productQuant)
-
-//         for (let index = 0; index < productQuant; index++) {
-//             if (index % 6 === 0) {
-//                 this.productsArray.pop(new Product(""))
-//             }
-//         }
-//         export default Catalog;
