--- src/modules/generateJoke.js
@@ -1,5 +1,0 @@
-function generateJoke() {
-    return "I dont trust stairs.they're always up to something."
-}
-
-export default generateJoke;
